fix: sqlite.rs file moved to src/persist/sqlite.rs
